Remove accidentally added language with wrong flag
I was confused since Duolingo used language code ha in that asset which is officially Hausa but in that case was Hawaiian, I checked and the Duo API uses hw, the correct code, in the response, so I simply deleted it.
There are a few other flag assets that duolingo has in their map file but don't have courses for yet but I will simply leave them here.